requirements.txt
tensorflow==1.9.0
Pillow==5.2.0
